requirements.txt
Django==1.9.8
djangorestframework==3.4.0
guess-language-spirit==0.5.2
six==1.10.0
